Changed return type of preprocessSHA() to be of type size_t.

+++ preprocess.h
@@ -4,6 +4,6 @@
 #include <stdio.h>
 #include <stdint.h>
 
-int preprocessSHA(FILE * stream, size_t msize, size_t csize);
+size_t preprocessSHA(FILE * stream, size_t msize, size_t csize);
 
 #endif
